[Python] Merge small PyROOT pythonization tests to cut process overhead
Each test in bindings/pyroot/pythonizations/test runs in its own Python
process, so every one pays ~2 s of interpreter startup + import ROOT
before a single assertion runs. For the many tiny per-class tests that
overhead is essentially the whole runtime.

Merge the small, overhead-bound tests by class family into single
modules (unittest discovery picks up all TestCase classes in a file).
This takes the directory from 78 to 51 registered tests with no loss of
coverage.
