--- bindings/pyroot/pythonizations/test/tfile_attrsyntax_get_writeobject_open.py
@@ -1,89 +1,0 @@
-import unittest
-
-import ROOT
-
-
-class TFileOpenReadWrite(unittest.TestCase):
-    """
-    Test for the TFile.Open factory like creation of TFile
-    """
-
-    filename  = 'tfileopenreadwrite.root'
-    nbins = 8
-    xmin = 0
-    xmax = 4
-
-    # Setup
-    @classmethod
-    def setUpClass(cls):
-        f = ROOT.TFile.Open(cls.filename, "RECREATE")
-        h = ROOT.TH1F("h", "h", cls.nbins, cls.xmin, cls.xmax)
-        ROOT.SetOwnership(h, False)
-        f.WriteObject(h, "h")
-
-        dir1 = f.mkdir("dir1")
-        dir1.cd()
-        h1 = ROOT.TH1F("h1", "h1", cls.nbins, cls.xmin, cls.xmax)
-        ROOT.SetOwnership(h1, False)
-        h1.Write()
-
-        dir2 = dir1.mkdir("dir2")
-        dir2.cd()
-        h2 = ROOT.TH1F("h2", "h2", cls.nbins, cls.xmin, cls.xmax)
-        ROOT.SetOwnership(h2, False)
-        h2.Write()
-
-        f.Close()
-
-    def checkHisto(self, h):
-        xaxis = h.GetXaxis()
-        self.assertEqual(self.nbins, h.GetNbinsX())
-        self.assertEqual(self.xmin, xaxis.GetXmin())
-        self.assertEqual(self.xmax, xaxis.GetXmax())
-
-    # Tests
-    def test_readHisto_itemsyntax(self):
-        f = ROOT.TFile.Open(self.filename)
-        self.checkHisto(f["h"])
-        self.checkHisto(f["dir1"]["h1"])
-        self.checkHisto(f["dir1"]["dir2"]["h2"])
-
-    def test_readHisto(self):
-        f = ROOT.TFile.Open(self.filename)
-        self.checkHisto(f.Get("h"))
-        self.checkHisto(f.Get("dir1/h1"))
-        self.checkHisto(f.Get("dir1/dir2/h2"))
-
-    def test_caching_getitem(self):
-        f = ROOT.TFile.Open(self.filename)
-        # check that object is not cached initially
-        self.assertFalse(hasattr(f, "_cached_items"))
-        f["h"]
-        # check that the value in __dict__ is actually the object
-        # inside the directory
-        self.assertTrue(f._cached_items['h'] is f["h"])
-
-    def test_oserror(self):
-        # check that an OSError is raised when an inexistent file is opened
-        # both with a string and an instance of TFileOpenHandle as arguments
-        self.assertRaises(OSError, ROOT.TFile.Open, 'inexistent_file.root')
-        handle = ROOT.TFile.AsyncOpen("inexistent_file.root")
-        self.assertRaises(OSError, ROOT.TFile.Open, handle)
-
-    def test_keys_title(self):
-        """
-        Test that the TKey related to a histogram in the file contains the
-        histogram title as described in #9989.
-        """
-        finput = ROOT.TFile.Open(self.filename)
-        key1 = finput.GetListOfKeys().At(0)
-        key2 = finput.Get("dir1").GetListOfKeys().At(0)
-        key3 = finput.Get("dir1/dir2").GetListOfKeys().At(0)
-        self.assertEqual(key1.GetTitle(), "h")
-        self.assertEqual(key2.GetTitle(), "h1")
-        self.assertEqual(key3.GetTitle(), "h2")
-        finput.Close()
-
-
-if __name__ == '__main__':
-    unittest.main()
