+++ bindings/pyroot/pythonizations/test/tobjstring.py
@@ -1,6 +1,39 @@
 import unittest
 
+import ROOT
 from ROOT import TObjString, TString
+
+
+class TObjStringLen(unittest.TestCase):
+    """
+    Test for the pythonization that provides the length of a
+    TObjString instance `s` via `len(s)`.
+    """
+
+    # Tests
+    def test_len(self):
+        s = 'test'
+        tos = ROOT.TObjString(s)
+        self.assertEqual(len(tos), len(s))
+
+
+class TObjStringStrRepr(unittest.TestCase):
+    """
+    Test for the pythonizations that provide a string representation
+    for instances of TObjString (__str__, __repr__).
+    """
+
+    # Tests
+    def test_str(self):
+        s = 'test'
+        tos = ROOT.TObjString(s)
+        self.assertEqual(str(tos), s)
+
+    def test_repr(self):
+        s = 'test'
+        tos = ROOT.TObjString(s)
+        self.assertEqual(repr(tos), repr(s))
+
 
 class TObjStringComparisonOps(unittest.TestCase):
     """
