+++ bindings/pyroot/pythonizations/test/tstring.py
@@ -1,6 +1,39 @@
 import unittest
 
+import ROOT
 from ROOT import TString
+
+
+class TStringLen(unittest.TestCase):
+    """
+    Test for the pythonization that provides the length of a
+    TString instance `s` via `len(s)`.
+    """
+
+    # Tests
+    def test_len(self):
+        s = 'test'
+        ts = ROOT.TString(s)
+        self.assertEqual(len(ts), len(s))
+
+
+class TStringStrRepr(unittest.TestCase):
+    """
+    Test for the pythonizations that provide a string representation
+    for instances of TString (__str__, __repr__).
+    """
+
+    # Tests
+    def test_str(self):
+        s = 'test'
+        ts = ROOT.TString(s)
+        self.assertEqual(str(ts), s)
+
+    def test_repr(self):
+        s = 'test'
+        ts = ROOT.TString(s)
+        self.assertEqual(repr(ts), repr(s))
+
 
 class TStringComparisonOps(unittest.TestCase):
     """
@@ -109,6 +142,48 @@ class TStringComparisonOps(unittest.TestCase):
         l2.sort()
 
         self.assertEqual(l1, l2)
+
+
+class TStringConverter(unittest.TestCase):
+    """
+    Tests for passing a Python string to a C++ function that expects a TString.
+
+    This feature is not implemented by a PyROOT pythonization, but by a converter
+    that was added to Cppyy to create a TString out of a Python string.
+    """
+
+    test_str = "test"
+
+    # Helpers
+    def check_type_conversion(self):
+        s = ROOT.TString(self.test_str)
+
+        # Works with TString...
+        self.assertEqual(ROOT.myfun(s), self.test_str)
+        # ... and Python string
+        self.assertEqual(ROOT.myfun(self.test_str), self.test_str)
+
+    # Tests
+    def test_by_value(self):
+        ROOT.gInterpreter.Declare("""
+        const char* myfun(TString s) { return s.Data(); }
+        """)
+
+        self.check_type_conversion()
+
+    def test_by_reference(self):
+        ROOT.gInterpreter.Declare("""
+        const char* myfun(TString &s) { return s.Data(); }
+        """)
+
+        self.check_type_conversion()
+
+    def test_by_const_reference(self):
+        ROOT.gInterpreter.Declare("""
+        const char* myfun(const TString &s) { return s.Data(); }
+        """)
+
+        self.check_type_conversion()
 
 
 if __name__ == '__main__':
