+++ bindings/pyroot/pythonizations/test/tvector3.py
@@ -3,6 +3,18 @@ import unittest
 import ROOT
 
 
+class TVector3Len(unittest.TestCase):
+    """
+    Test for the pythonization that allows to get the size of a
+    TVector3 (always 3) by calling `len` on it.
+    """
+
+    # Tests
+    def test_len(self):
+        v = ROOT.TVector3(1., 2., 3.)
+        self.assertEqual(len(v), 3)
+
+
 class TVector3GetItem(unittest.TestCase):
     """
     Test for the pythonization that allows to: (i) get an item of a
