+++ bindings/pyroot/pythonizations/test/tvectort.py
@@ -3,6 +3,20 @@ import unittest
 import ROOT
 
 
+class TVectorTLen(unittest.TestCase):
+    """
+    Test for the pythonization that allows to get the size of a
+    TVectorT by calling `len` on it.
+    """
+
+    num_elems = 3
+
+    # Tests
+    def test_len(self):
+        v = ROOT.TVectorT[float](self.num_elems)
+        self.assertEqual(len(v), self.num_elems)
+
+
 class TVectorTGetItem(unittest.TestCase):
     """
     Test for the pythonization that allows to: (i) get an item of a
